refactor: "입력 클래스, 출력 클래스 리팩토링" - 입력클래스   - Question 메시지 상수 위임   - 메시지 상수에 맞는 메서드 생성 - 출력클래스   - startLine만 "실행 결과" 출력 추가
